+++ R/databases.R
@@ -799,10 +799,12 @@ updateGrant <- function(databaseId, userId, resourceId, permissions) {
 #' @examples
 #' \dontrun{
 #'
-#' # Use the current recommended grant-based roles 
+#' # Use the current grant-based roles; legacy roles are deprecated
 #' grantBased = TRUE
+#' dbId = "cxy123"
 #'
 #' if (grantBased) {
+#'   
 #'   currentGrantBasedRole <- 
 #'     role(id = "rp",
 #'         label = "Reporting Partner",
@@ -822,7 +824,24 @@ updateGrant <- function(databaseId, userId, resourceId, permissions) {
 #'               add_record = TRUE),
 #'             optional = TRUE))
 #'         )
-#'   updateRole("cxy123", currentGrantBasedRole) 
+#'
+#'   # Duplicate the role with a different id
+#'   currentGrantBasedRole2 <- currentGrantBasedRole
+#'   currentGrantBasedRole2$id <- "rp2"
+#'
+#'   addRole(dbId, currentGrantBasedRole)
+#'   addRole(dbId, currentGrantBasedRole2)
+#'   
+#'   currentGrantBasedRole$label <- "Original reporting orgs"
+#'   updateRole(dbId, currentGrantBasedRole)
+#'   
+#'   deleteRoles(dbId, c(currentGrantBasedRole$id,currentGrantBasedRole2$id))
+#'   
+#'   # delete all roles containing "readonly" - will fail if assigned to a user
+#'   remainingRoles <- sapply((getDatabaseTree(dbId))$roles, function(x) x$id)
+#'   readOnlyRoles <- remainingRoles[grepl("readonly", remainingRoles)]
+#'   deleteRoles(dbId, roleIds = readOnlyRoles)
+#'    
 #' } else {
 #'   # These older-style roles will be phased out.
 #'   deprecatedNonGrantRole <- list(
@@ -884,6 +903,8 @@ addRole <- function(databaseId, role) {
   }
 }
 
+#' @param roleIds the ids of the roles to be deleted. It should be passed as a character vector.
+#'
 #' @rdname updateRole
 #' @order 3
 #' @export
@@ -896,7 +917,7 @@ deleteRoles <- function(databaseId, roleIds) {
   request <- databaseUpdates()
   request$roleDeletions = lapply(roleIds, function(x) x)
   
-  x <- postResource(path, request, task = "updateRole")
+  x <- postResource(path, request, task = "deleteRoles")
   invisible()
 }